requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.7
Werkzeug==2.0.2
gunicorn==20.1.0
Pillow==8.4.0